draft

+++ src/jsp/com/cubrid/jsp/data/DBType.java
@@ -31,6 +31,14 @@
 
 package com.cubrid.jsp.data;
 
+import com.cubrid.jsp.value.DoubleValue;
+import com.cubrid.jsp.value.FloatValue;
+import com.cubrid.jsp.value.IntValue;
+import com.cubrid.jsp.value.LongValue;
+import com.cubrid.jsp.value.NullValue;
+import com.cubrid.jsp.value.ShortValue;
+import com.cubrid.jsp.value.StringValue;
+import com.cubrid.jsp.value.Value;
 import cubrid.jdbc.driver.CUBRIDBinaryString;
 import cubrid.jdbc.driver.CUBRIDBlob;
 import cubrid.jdbc.driver.CUBRIDClob;
@@ -75,7 +83,7 @@ public class DBType {
     public static final int DB_DATETIMETZ = 38;
     public static final int DB_DATETIMELTZ = 39;
 
-    public static byte getObjectDBtype(Object value) {
+    public static int getObjectDBtype(Object value) {
         if (value == null) return DB_NULL;
         else if (value instanceof String) return DB_STRING;
         else if (value instanceof Byte) return DB_SHORT;
